# Optimized Bash implementation