--- a/PrimeBash/solution_1/Dockerfile
+++ b/PrimeBash/solution_1/Dockerfile
@@ -1,8 +1,8 @@
-FROM ubuntu:20.04
-
-RUN apt-get update
+# bash (with coreutils) in alpine seems to be less optimized than ubuntu
+# bash 5.1 seems to have much better performance than 5.0 shipped in 20.04
+FROM ubuntu:21.04
 
 WORKDIR /opt/app
-COPY PrimeBash.sh .
-
-ENTRYPOINT [ "./PrimeBash.sh" ]
+COPY *.sh ./
+ 
+ENTRYPOINT [ "./run.sh" ]