--- a/PrimeBash/solution_1/run.sh
+++ b/PrimeBash/solution_1/run.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cd "${0%/*}"
+
+exec find . -name '*.sh' -and -not -name run.sh -exec {} \;