## Description
I'm not entirely sure if using properties of uninitialized values is considered legal, but if so these changes do a little more than double performance of solution_1 on my machine (~12 seconds to ~5 seconds).  Solution_2 included here uses the same trick as PowerShell's solution_2 but only takes the run time for me from ~5 seconds to ~3 seconds.

As an aside I did try bitpacking to avoid the initialization slowness, but it seems that Bash's performance in mathematical code is basically a function of the number of variable accesses that are being done.  Doing the math, and creating the local variables required, to bit pack basically costs as much as it saves vs not bit packing.

I did try implementing a faster sqrt algorithm but once again the complexity of doing a smarter algorithm seemed to be outweighed by overheads on executing commands and variable access.  Plus it would only matter for smaller sieveSizes which the simple algorithm doesn't take long with anyway.  So I got rid of the subshell and called it a day.

Oh I saw a note somewhere that the reason Ubuntu was used for the container instead of Alpine was because the script wasn't exiting on Alpine.  This was likely due to the lack of coreutils for the date command, but even fixing that I found Alpine's Bash to be about half the speed of Ubuntu's.  Not sure why, but have to assume libc differences?

In summary the changes are:
- Invert meaning of 0 and 1 to avoid need for slow initialization. (Uninitialized values act as 0 for the purposes of this script.)
- Reduce complexity of get and clear functions for significant performance improvement.
- Use associative arrays for bitArray as they're actually faster in Bash (possibly due to avoiding the expression evaluator on the index)
- Upgrade to Bash 5.1 (Ubuntu 21.04) for improved performance. Note, Bash 5.1 in Alpine is about half the speed.
- Use a different method of limiting the runtime of the script which doesn't require spawning a process each loop.
- Removed the need for a subshell for the sqrt call.
- Set LC_ALL to C to avoid localization costs.
- Fixed typo on printing of times (9 digits instead of 8).
- Wrapped main in a function so that some of the parsing is done ahead of time instead of line by line.
- Added solution 2 which does the same inlining as PowerShell's solution 2 (albeit Bash's function call overhead isn't anywhere near as bad)

## Contributing requirements

* [x] I read the contribution guidelines in CONTRIBUTING.md.
* [x] I placed my solution in the correct solution folder.
* [x] I added a README.md with the right badge(s).
* [x] I added a Dockerfile that builds and runs my solution.
* [x] I selected `drag-race` as the target branch.
* [x] All code herein is licensed compatible with BSD-3.