@rbergen 2021-07-12
Please convert solution_2 to a second implementation of solution_1. It will save docker build time and resources during the benchmarks. Also already tagging @Nitepone so they can provide feedback on this PR on their original implementation, if they want to do so.