@Nitepone 2021-07-12
These look like excellent changes to me! I had not spent much time optimizing my original implementation. I also have never even tried to optimize bash code for time. (I also hadn't much time to even fix my original PR last month. Thanks @rbergen for getting my original implementation merged in!)

A minor suggestion would be to remove the getBit and clearBit functions from the solution_2 solution. clearBit is dead code and getBit could be easily replaced (alas, for no test time performance benefit).

Also, you could update the readme/file to accredit your changes.

Awesome PR @Blzut3, I didn't expect to see a 4-5x improvement when running solution_2 on my machine! Made my morning to see this code get much needed improvements! :tada: