@Blzut3 2021-07-13
@rbergen The contributing instructions appear to use the term solution and implementation interchangeably so I'm not sure what your request is supposed to look like as far as the execution framework goes?  Is there an example from another language that I can reference?  My best guess is you want me to include both run runSieve functions in one file and run both main loops back to back, but not sure.