@rbergen 2021-07-13
@Blzut3 PrimeBASIC/solution_1 demonstrates what my request aims for. It's one solution directory with one Dockerfile, that (in the BASIC case compiles and) runs multiple implementations (approaches) in one Docker image. As that solution also shows, it is not necessary to include both implementations in one file, as long as they are executed within the one Docker image.