@Blzut3 2021-07-14
@Nitepone Took out the clearBit function, but left getBit in since as you say it would only affect code outside of the drag race.  Makes things slightly easier to change/experiment with if the other code is all the same.

@rbergen Done.  Also added the bit packed implementation as a separate commit.  Feel free to toss it out if you wish, but figured perhaps someone would be curious as to why it wasn't done, and it's just a matter of dropping in a different getBit/clearBit function pair.

Also, kind of curious what it would take for an implementation to be considered faithful?  Is the issue just that it uses global variables for the state instead of passing it through the functions?  Kind of feel like the rules are technically met by emptyBitArray tossing the old state (possibly more technically correct if there was an unset first?), but I can definitely see how such would be twisting the rules.  Creating a unique variable name for bitArray each iteration and passing it to the inlined runSieve wouldn't be particularly difficult and wouldn't cost much in runtime performance.  But I don't know if one would need to go as far as generating unique variable names, or if it's just the act of passing the variable as an argument that's enough to do it?