@rbergen 2021-07-14
> Also, kind of curious what it would take for an implementation to be considered faithful?

If I'm honest, I do think the second rule in the "faithfulness" section of the contributing guidelines does answer your question, but as I'm typing anyway: what's primarily missing here is a "class-like" wrapper object that encapsulates the full sieve state, that being the prime candidate flags and whatever else the implementation needs to run the sieve. That wrapper object needs to be created (and torn down in languages where that is a developer action) for every sieve run,