@Nitepone 2021-07-14
> what's primarily missing here is a "class-like" wrapper object that encapsulates the full sieve state, that being the prime candidate flags and whatever else the implementation needs to run the sieve. That wrapper object needs to be created (and torn down in languages where that is a developer action) for every sieve run,

Bash does not contain a class construct. The contributing guidelines specify that such languages should use the "closest thing" to such functionality. This sadly leaves some ambiguity in this case. 😥

```
It uses a class to encapsulate the sieve, or a (closest) equivalent feature in your language if a class construct is not available. 
```

Due to the nature of Bash, globals variables are often required. (i.e. Returning an array from a bash function).

Creating class-like wrappers is not something I've ever seen in the wild. It is common to maintain globals through functions though. By accessing the globals entirely through functions, you can achieve a (weak) level of encapsulation.