@rbergen 2021-07-14
If I'm honest, my thinking is that there is no way to create a faithful implementation in Bash. As it has not been possible to do so in quite a few other purely procedural languages. It's important to keep in mind that in the context of this repo "not faithful" does not constitute a value statement. It's just a statement on one of the four key characteristics of an implementation.