@Nitepone 2021-07-14
That makes sense, and is what I expected. It's a shame, but bash just isn't the best language to use to find prime numbers.